# Reducing list API payload