### Description
Fetching only selected attributes using [query.SetAttrSelection](https://github.com/googleapis/google-cloud-go/blob/storage/v1.45.0/storage/storage.go#L1784) for ListObjects API which are sufficient for serving the request and reduce the payload for network call.
Also included :  necessary changes in the fake bucket tests.


Perf Metrics for listing : 
| Test Directory Size ( in number of files) | Mean Latency(Master) (in msec) | Mean Latency(reduced payload) (in msec) | Percentage Improvement |
| ------------------- | ---------------------- | --------------------------------- | ------------------------- |  
| 1000 | 61.153  | 56.325| 7.89% |
| 10000 | 462.775 | 410.701 | 11.25%|
| 50000 |  2373.293 |2103.235 | 11.38%|
| 100000 | 4677.908 | 4354.92| 6.90% |
| 200000 |9417.683 |8628.266 | 8.38% |

Note: More detailed metrics [here](https://docs.google.com/spreadsheets/d/101rbvNWjU55sTUaNP9p86D_g1A_2vxCjq7FSmyPDuzo/edit?usp=sharing)
### Link to the issue in case of a bug fix.
NA

### Testing details
1. Manual - NA
2. Unit tests - NA
3. Integration tests - NA